+++ ceammc/CHANGELOG.md
@@ -1,4 +1,8 @@
 # CEAMMC extension library changelog
 
-## Unreleased
+## [Unreleased]
+### Fixed:
+- array.phase->sample fix with lists argument
+- array.phase->sample clip phase to [0, 1] range
+- array.phase->sample help file added